backend/requirements.txt
# Install with Aliyun mirror (recommended in China):
#   pip install -r requirements.txt -i https://mirrors.aliyun.com/pypi/simple/

fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langgraph>=0.2.0
langchain-core>=0.3.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
PyJWT>=2.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
bcrypt>=4.0.0
